requirements.txt
torch
sentence-transformers
haystack-ai
datasets
gradio
openai
pyyaml
